Backend/requirements.txt
envDjango
PyQRCode 
pypng